refactor: move utils to own module
